+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Added `Otaku` Background.
+- Added `Reborn Demon Lord` Background.
 
 ## [1.6.0] - 2022-11-07
 ### Added